+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [v0.2.16]
 - Minor text corrections
 - Support wasi2ic as a library
-
+- Add command line option to print out existing exports
+- After conversion, report if there are any unresolved dependencies left, return error code 1 if there is anything other than ic0
+- Update dependencies
 
 ## [v0.2.15]
 - Update dependencies